Reworking threads a little bit.
Looks like only the tick event uses a thread to find blocks while
pickupItem, placeItem, and config call blockFinder directly.
This may lead to the method called twice in parallel
(which may have caused issues before the delay was introduced).

Part of the rework:
  * forbid calls to blockFinder(), force the use of XRay.requestBlockFinder()
    which handles threads
  * use a singleThreadExecutor
  * make sure the task only runs once at a time with no queue
    (subsequent calls are ignored or they'll stack up quickly)
  * let threads run when they can (no need for a delay anymore)
  * secure concurrent access to shared objects (XRayRenderer.ores and
    XRay.drawOres for instance)

Current status: work in progress :)
It compiles, it works, no dead-lock found. It requires a lot of
testings though.